+++ summarynb/__init__.py
@@ -6,7 +6,7 @@ from IPython.display import HTML, display
 
 __author__ = """Maxim Zaslavsky"""
 __email__ = "maxim@maximz.com"
-__version__ = '0.1.3'
+__version__ = "0.1.4"
 
 """Functions that return functions that return HTML."""
 
@@ -22,11 +22,15 @@ def image(img_src):
     :rtype: function
     """
 
-    def template(max_width, max_height):
+    # Convert absolute path to relative path, since a browser won't be able to grab an image from "/home/..."
+    img_src = os.path.relpath(img_src)
+
+    def template(max_width, max_height, *args, **kwargs):
         def convert_to_px_or_unset(optional_numeric_value):
             if not optional_numeric_value:
-                return 'inherit'
-            return str(optional_numeric_value) + 'px'
+                return "inherit"
+            return str(optional_numeric_value) + "px"
+
         max_width = convert_to_px_or_unset(max_width)
         max_height = convert_to_px_or_unset(max_height)
 
@@ -46,8 +50,36 @@ def table(df):
     :rtype: function
     """
 
-    def template(max_width, max_height):
+    def template(*args, **kwargs):
         return df.to_html()
+
+    return template
+
+
+def plaintext(text, **kwargs):
+    """Renders plain text.
+
+    :param text: text to display
+    :type text: str
+    """
+
+    def template(*args, **kwargs):
+        # convert to html
+        return f"<pre>{text}</pre>"
+
+    return template
+
+
+def empty(width=None):
+    """Creates empty separator block.
+
+    :param width: specify block width, defaults to None
+    :type width: int, optional
+    """
+
+    def template(max_width, *args, **kwargs):
+        # empty cell
+        return f'<div style="min-width: {max_width if not width else width}px;"></div>'
 
     return template
 
@@ -73,6 +105,15 @@ def indexed_csv(fname, cols=None, **kwargs):
     return csv(fname, cols=cols, index_col=0, **kwargs)
 
 
+def textfile(fname, **kwargs):
+    """
+    Read a text file and render as plain text.
+    """
+    with open(fname, "r") as f:
+        text = f.read()
+        return plaintext(text)
+
+
 """Main logic."""
 
 
@@ -89,8 +130,10 @@ def _get_template(user_input):
     # detect different table types
     if extension == ".csv":
         return csv(user_input)
-    if extension in [".tsv", ".txt"]:
+    elif extension == ".tsv":
         return csv(user_input, sep="\t")
+    elif extension == ".txt":
+        return textfile(user_input)
 
     # assume it's an image
     return image(user_input)
@@ -132,8 +175,8 @@ def _flatten(lst):
 
 def chunks(entries, shape):
     """Reshape [entries] into chunks of shape [shape], which can be:
-        - a (number of rows, number of columns) tuple, in which case we verify length
-        - or simply a number of columns, in which case we guess the right number of rows, and allow in-complete rows."""
+    - a (number of rows, number of columns) tuple, in which case we verify length
+    - or simply a number of columns, in which case we guess the right number of rows, and allow in-complete rows."""
     # Flatten entries
     entries = _flatten(entries)
 
@@ -197,7 +240,10 @@ def _make_HTML(entries, headers, max_width, max_height):
     # Make HTML for each row
     rows = [
         "\n".join(
-            [wrap_in_column(_get_template(template)(max_width, max_height)) for template in row]
+            [
+                wrap_in_column(_get_template(template)(max_width, max_height))
+                for template in row
+            ]
         )
         for row in entries
     ]
@@ -224,4 +270,10 @@ def show(entries, headers=None, max_width=800, max_height=800):
         - [max_height]: set max pixel height for images, default 800px. set to None to disable max height.
     """
 
-    return display(HTML(_make_HTML(entries, headers=headers, max_width=max_width, max_height=max_height)))
+    return display(
+        HTML(
+            _make_HTML(
+                entries, headers=headers, max_width=max_width, max_height=max_height
+            )
+        )
+    )