+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 0.1.4
+
+* Bugfix: always provide image URLs as relative paths, even if user passes an absolute path
+* Pre-commit hook support
+* Plaintext and empty-block handlers
+## 0.1.3
+
+* Better max-width and max-height dimension configurability
+## 0.1.2
+
+* Clearer readme.
+## 0.1.1
+
+* Clearer readme.
+
 ## 0.1.0
 
 * First release on PyPI.